Enforce layered sync architecture: SwiftSync = pure storage, engine = orchestration
Reverts the outbound fold into SyncContainer (it produced a mixed-@MainActor
smell — orchestration is networking-layer policy, not storage):
- SyncContainer: drop SyncBackend/register/drain/isOnline/onDrainComplete; back
  to a pure @unchecked Sendable storage primitive.
- DemoSyncEngine: reabsorbs the drain orchestration — calls
  SwiftSync.withPendingChanges with its own upload (networking), coalesces
  concurrent pushes, drains on reconnect (isOffline didSet), and keeps the
  inbox-safety (a thrown drain doesn't re-annotate).
- Docs: new docs/project/architecture.md (the four layers, SwiftSync's role,
  generics, data-from-store-vs-status-from-machine); sync-engine-fold.md marked
  superseded; README's offline section points to it; roadmap updated.

SwiftSync 171 / DemoCore 42 / DemoBackend 30 green; demo app builds.
